move 'stored' related filters to separate filter dropdown
