+++ main/res/values/strings.xml
@@ -984,6 +984,7 @@
     <string name="ignore_confirm_title">Be careful</string>
     <string name="ignore_confirm_message">Ignoring a cache means that this cache will never occur again when loading data from geocaching.com. You will only be able to see the cache again by unignoring it on the website.</string>
     <string name="cache_status">Status</string>
+    <string name="cache_stored">Stored</string>
     <string name="cache_status_offline_log">Saved Log</string>
     <string name="cache_status_found">Found</string>
     <string name="cache_not_status_found">Not Found</string>
@@ -995,14 +996,14 @@
     <string name="cache_status_not_premium">All Members Access</string>
     <string name="cache_status_stored">Stored</string>
     <string name="cache_status_not_stored">Not stored</string>
-    <string name="cache_status_stored_lasteighthours">Stored within a) last eight hours</string>
-    <string name="cache_status_stored_lasttwentyfourhours">Stored within b) last twenty-four hours</string>
-    <string name="cache_status_stored_lastsevendays">Stored within c) last seven days</string>
-    <string name="cache_status_stored_lastthirtydays">Stored within d) last thirty days</string>
-    <string name="cache_status_stored_morethaneighthours">Stored before a) last eight hours</string>
-    <string name="cache_status_stored_morethantwentyfourhours">Stored before b) last twenty-four hours</string>
-    <string name="cache_status_stored_morethansevendays">Stored before c) last seven days</string>
-    <string name="cache_status_stored_morethanthirtydays">Stored before d) last thirty days</string>
+    <string name="cache_status_stored_lasteighthours">Within last eight hours</string>
+    <string name="cache_status_stored_lasttwentyfourhours">Within last twenty-four hours</string>
+    <string name="cache_status_stored_lastsevendays">Within last seven days</string>
+    <string name="cache_status_stored_lastthirtydays">Within last thirty days</string>
+    <string name="cache_status_stored_morethaneighthours">Before last eight hours</string>
+    <string name="cache_status_stored_morethantwentyfourhours">Before last twenty-four hours</string>
+    <string name="cache_status_stored_morethansevendays">Before last seven days</string>
+    <string name="cache_status_stored_morethanthirtydays">Before last thirty days</string>
     <string name="cache_geocode">Geo code</string>
     <string name="cache_name">Name</string>
     <string name="cache_type">Type</string>
